+++ decidim-proposals/app/controllers/decidim/proposals/admin/proposals_merges_controller.rb
@@ -16,7 +16,7 @@ module Decidim
             end
 
             on(:invalid) do
-              flash.now[:alert] = I18n.t("proposals_merges.create.invalid", scope: "decidim.proposals.admin")
+              flash[:alert] = I18n.t("proposals_merges.create.invalid", scope: "decidim.proposals.admin")
               redirect_to EngineRouter.admin_proxy(current_component).root_path
             end
           end